Create FICE.cpp
The number of ways a given number can be expressed as the sum of odd numbers is actually the Fibonacci result of that number.
e.g 
4 can be expressed in 3 ways 
i.e 1+1+1+1, 1+3, 3+1
fib(4)=3
similarly :-1: 
fib(1)=1 
fib(2)=1              1+1
fib(3)=2              1+1+1 ,3
fib(5)=8              1+1+1+1+1,1+1+3,1+3+1,3+1+1,5

Fibonacci can be calculated in many ways ,
the above shown method is an efficient way of calculating fibonacci with time complexity of O(logn).

It is a proven fact that if a matrix F={{1,1},{1,0}} is self multiplied n times then the resultant F[0][0] will be the fibonacci number.

But this question requires calculation of 1000th fibonacci number which is extremely huge for the memory to hold.
If you read carefully, it has actually asked us to calculate the modulus of twice of the resulting fibonacci number. Now this can be calculated easily if we keep on taking modulus while calculating the value of F[0][0]. This way the values remain small in size.