Create CHEFST.cpp
hint: if the sum of series 1...m is less than the minimum of the n1 and n2 then all you have to do n1-sum(1..m) + n2-sum(1..m).
else if the sum(1..m)>min(n1,n2)
then it seems obvious that minimum of n1 and n2 would reduce down to zero at some point.
e.g 
n1   n2   m
10   8     3
7     5     2
5     3     1
4     2     0

We see that 8 has reduced down to 2 but not zero because 3+2+1+0=6 which is less than 8.

e.g 2
n1    n2    m
10    8      5
5      3      3
2      0      0

We see that 5+4+3+2+1=15 which is greater than 8. Therefore we know that after some subtractions it will come down to zero. Then why waste computations in doing those subtractions? Just subtract both n1 and n2 (10 and 8) by 8. Which would give 2 and 0.

Q: Why did you subtract max(n1,n2)-min(n1,n2) to calculate d and what is d+2*n?
A:  It's simple. A+B = A-B+2B.
     d=A-B;  and n=B;
     Also, we are subtracting both n1 and n2 with the same number , so the difference will still remain the same.