Logic
Hint: Sort the pairs in ascending order. Then remove all the brand duplicates. This would give a vector where all the brands would be different and hence as we go down the array the 'no. of different brands'  will increase. 

The duplicates brand's length can be directly multiplied with the no. of distinct brands in the above vector. Because obviously, after all the distinct brands are played, the 'no.of different brands' won't change.